+++ R/boostrap_clust_validation.R
@@ -221,7 +221,7 @@ boot_group_validation <- function(object, nsim = 1000, maxiter = 1000, tol = 0.0
 #' Cmean <- CMeans(Data,4,1.5,500,standardize = FALSE, seed = 456,
 #'     tol = 0.00001, verbose = FALSE)
 #'
-#' future::plan(future::multiprocess(workers=2))
+#' future::plan(future::multisession(workers=2))
 #'
 #' validation <- boot_group_validation.mc(Cmean, nsim = 1000, maxiter = 1000,
 #'     tol = 0.01, init = "random")
@@ -353,7 +353,7 @@ boot_group_validation.mc <- function(object, nsim = 1000, maxiter = 1000, tol = 
 }
 
 
-#' @title Bootstrap check the robustness of a classification (multicore)
+#' @title Worker function for cluster bootstrapping
 #'
 #' @description Worker function for cluster bootstrapping
 #'